refactor: Restaurant 엔티티와 관련 enum 및 embedded 타입을 별도의 도메인 패키지로 이동
