+++ src/main/java/com/jvnlee/catchdining/domain/restaurant/model/CountryType.java
@@ -1,4 +1,4 @@
-package com.jvnlee.catchdining.entity;
+package com.jvnlee.catchdining.domain.restaurant.model;
 
 public enum CountryType {
     KOREAN, JAPANESE, CHINESE, AMERICAN, ITALIAN, FRENCH, SPANISH, FUSION, ETC
