+++ src/main/java/com/jvnlee/catchdining/domain/restaurant/model/FoodType.java
@@ -1,4 +1,4 @@
-package com.jvnlee.catchdining.entity;
+package com.jvnlee.catchdining.domain.restaurant.model;
 
 public enum FoodType {
     BEEF, PORK, LAMB, CHICKEN, DUCK, SEAFOOD, SUSHI, VEGAN, ETC
