+++ src/main/java/com/jvnlee/catchdining/domain/restaurant/model/ServingType.java
@@ -1,4 +1,4 @@
-package com.jvnlee.catchdining.entity;
+package com.jvnlee.catchdining.domain.restaurant.model;
 
 public enum ServingType {
     OMAKASE, CUISINE, COURSE, BUFFET, BARBEQUE
